Split RegistryTests into those that need 'docker' and those that do not.
